requirements.txt
aws-cdk-lib>=2.211.0
constructs>=10.4.2
cdk-nag>=2.37.6
boto3>=1.40.12
jsii>=1.113.0
